Stat Component update
Added the logic to handle when the character's satiation or hydration drop below a certain threshold to drain stamina and drain health. The actual thresholds will be added later. Also added the blood stat. The blood stat will dictate whether the player's health is able to regenerate or not. Will add the logic at a later point.